requirements.txt
locustio==0.14.5
assertpy==1.0